pyproject.toml
[tool.poetry]
name = "reqres-api-tests"
version = "0.1.0"
description = ""
authors = ["Evgenii Li <evgeniili2203@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
jsonschema = "^4.19.1"
pytest = "^7.4.2"
allure-pytest = "^2.13.2"
curlify = "^2.2.1"

[tool.pytest.ini_options]
addopts = "--clean-alluredir --alluredir=allure-results"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
